requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Flask-Ext==0.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2

click==8.0.4
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.0.3

gspread==3.1.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
oauth2client==4.1.3
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dotenv==0.21.0
requests>=2.22.0
rsa==4.0
six==1.12.0
spotipy==2.4.4
SQLAlchemy==1.3.5
uritemplate==3.0.0
urllib3>=1.25.3
WTForms==2.2.1
